requirements.txt
acme==2.6.0
asgiref==3.6.0
CacheControl==0.12.11
cachetools==5.3.0
certbot==2.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
ConfigArgParse==1.5.3
configobj==5.0.8
cryptography==40.0.2
cssselect==1.2.0
Deprecated==1.2.13
distro==1.8.0
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
firebase-admin==6.1.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
josepy==1.13.0
jwcrypto==1.4.2
lxml==4.9.2
msgpack==1.0.5
oauth2client==4.1.3
parsedatetime==2.6
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
Pyrebase4==4.7.0
pyRFC3339==1.1
python-jwt==4.0.0
pytz==2023.3
pywin32==306
requests==2.30.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
